requirements.txt
Django>=5.2.9
djangorestframework
dj-rest-auth
django-cors-headers
djangorestframework-simplejwt
Pillow
